HelpUsImproveHoneyModal: rename and restructure
