--- src/modals/HelpUsImproveHoney/index.js
@@ -1,3 +1,0 @@
-import HelpUsImproveHoney from './HelpUsImproveHoney.container'
-
-export default HelpUsImproveHoney
