+++ src/modals/HelpUsImproveHoneyModal/HelpUsImproveHoneyModal.container.js
@@ -1,9 +1,10 @@
 import { connect } from 'react-redux'
 
 import UIActions from '../../redux/actions/ui'
+import WSActions from '../../redux/actions/ws'
 import { getUIModalStateForKey, getThemeSetting } from '../../redux/selectors/ui'
 // import { getAuthToken } from '../../redux/selectors/ws'
-import HelpUsImproveHoney from './HelpUsImproveHoney'
+import HelpUsImproveHoneyModal from './HelpUsImproveHoneyModal'
 import { UI_MODAL_KEYS } from '../../redux/constants/modals'
 
 const mapStateToProps = (state = {}) => ({
@@ -13,9 +14,10 @@ const mapStateToProps = (state = {}) => ({
 })
 
 const mapDispatchToProps = dispatch => ({
+  updateSettings: (key, value) => dispatch(WSActions.saveSettings(key, value)),
   closeHelpUsImproveHoneyModal: () => {
     dispatch(UIActions.changeUIModalState(UI_MODAL_KEYS.HELP_US_IMPROVE_HONEY_MODAL, false))
   },
 })
 
-export default connect(mapStateToProps, mapDispatchToProps)(HelpUsImproveHoney)
+export default connect(mapStateToProps, mapDispatchToProps)(HelpUsImproveHoneyModal)
