+++ src/modals/HelpUsImproveHoneyModal/index.js
@@ -1,0 +1,3 @@
+import HelpUsImproveHoneyModal from './HelpUsImproveHoneyModal.container'
+
+export default HelpUsImproveHoneyModal
