+++ src/modals/ModalsWrapper/ModalsWrapper.js
@@ -12,7 +12,7 @@ const ClosePositionModal = lazy(() => import('../ClosePositionModal'))
 const ConfirmDMSModal = lazy(() => import('../ConfirmDMSModal'))
 const EditOrderModal = lazy(() => import('../EditOrderModal'))
 const ResetAPIKeysModal = lazy(() => import('../ResetAPIKeysModal'))
-const HelpUsImproveHoney = lazy(() => import('../HelpUsImproveHoney'))
+const HelpUsImproveHoneyModal = lazy(() => import('../HelpUsImproveHoneyModal'))
 
 const ModalsWrapper = ({ isElectronApp }) => {
   return (
@@ -25,13 +25,13 @@ const ModalsWrapper = ({ isElectronApp }) => {
           <AppSettingsModal />
           <CloseSessionModal />
           <ResetAPIKeysModal />
+          <HelpUsImproveHoneyModal />
         </>
       )}
       <NoConnectionActionModal />
       <BadConnectionModal />
       <CcyInfoModal />
       <EditOrderModal />
-      <HelpUsImproveHoney />
       <ClosePositionModal />
     </>
   )
